Blocks: Move components to subdirectory
