+++ lib/blocks/src/components/BlockBackgroundStyles.tsx
