+++ lib/blocks/src/components/ColorPalette.stories.tsx
