+++ lib/blocks/src/components/ColorPalette.tsx
