+++ lib/blocks/src/components/DocsPage.tsx
