+++ lib/blocks/src/components/DocsPageExampleCaption.md
