+++ lib/blocks/src/components/DocsPageExampleCaption.mdx
