+++ lib/blocks/src/components/Story.tsx
