+++ lib/blocks/src/components/index.ts
@@ -1,0 +1,11 @@
+export * from './Source';
+export * from './EmptyBlock';
+export * from './Description';
+export * from './DocsPage';
+export * from './Preview';
+export * from './ArgsTable';
+export * from './Story';
+export * from './IFrame';
+export * from './Typeset';
+export * from './ColorPalette';
+export * from './IconGallery';
