+++ lib/blocks/src/index.ts
@@ -1,14 +1,2 @@
-export * from './Source';
-export * from './EmptyBlock';
-export * from './Description';
-export * from './DocsPage';
-export * from './Preview';
-export * from './ArgsTable';
-export * from './Story';
-export * from './IFrame';
-export * from './Typeset';
-export * from './ColorPalette';
-export * from './IconGallery';
-
-// Controls
+export * from './components';
 export * from './controls';
